New API using indexes to access students and evals
Now, one cannotuse the Students and Evals objects anymore. The new Tgc's
functions use an index to access these objects and their methods.
New find functions and iterators have been added to find these indexes.

Evaluations category rules are not finished and there is still no way to get
a report (work in progress).

Competencies are not handled for now...